requirements.txt
# ── Windows / Linux ──────────────────────────────────────────────────────────
# Install: pip install -r requirements.txt

# Adafruit / GPIO (no RPi-specific packages on Windows/Linux)
Adafruit-Blinka==8.69.0
adafruit-circuitpython-dht==4.0.10

# Google APIs
google-genai==1.68.0
supabase==2.18.1

# Encryption
cryptography>=41.0.0

# Computer Vision
opencv-python==4.13.0.92

# Numerical
numpy==2.4.2

# Environment
python-dotenv==1.2.1
